GeoCities WayBack Machine Archive.org capture of eTude316 TL and TL Construction plans
Finding these was time consuming and difficult. Posting a copy here with attrubtion to the author. Paul Spenser of Red Spade Audio. 

This website [ https://www.diyaudio.com/community/threads/main-speaker-kits-which-ones.24999/ ] links to the GeoCities page for Paul's plans for the TL box for the JayCar JV60 which had something of a cult following in Australia back in the mid/late 1990s and early 2000s. The speaker was produced and sold by Jaycar until the early 2010s in kit form. The odd kit is still available to be had if you are lucky and keep searching. Only took me 3 years! 

Way Back Machine internet archive.org capture of the respective pages before GeoCities was shutdown by Yahoo.
1) http://web.archive.org/web/20091027144207/http://www.geocities.com/etude316/audio/TL.html
2) http://web.archive.org/web/20091027144211/http://geocities.com/etude316/audio/TLconstruction.html
3) Red Spade Audio on Facebook (as of 2023): https://www.facebook.com/RedSpadeAudio/